TFM/cloud-functions/cv-pipeline/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.34.0
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.100.0
google-cloud-logging==3.12.1
pydantic==2.11.5